Update to version 1.0.10.1
New Clipper library added with some small fixes.
Group Layers now supported.
Depth buffer fixes added.

+++ unity/Tiled2Unity/Scripts/Editor/Tiled2UnityMenuItems.cs
@@ -32,7 +32,9 @@ namespace Tiled2Unity
             if (path.Length != 0)
             {
                 List<string> packageFiles = new List<string>();
-                packageFiles.AddRange(EnumerateAssetFilesAt("Assets/Tiled2Unity", ".cs", ".shader", ".txt"));
+
+                // Export all C# files, shaders, text files, and some select materials
+                packageFiles.AddRange(EnumerateAssetFilesAt("Assets/Tiled2Unity",".cs", ".shader", ".txt", "t2uSprite-Depth.mat", "t2uSprite-DiffuseDepth.mat"));
                 AssetDatabase.ExportPackage(packageFiles.ToArray(), path);
             }
         }
@@ -49,11 +51,11 @@ namespace Tiled2Unity
         //    DeleteAssetsAt("Assets/Tiled2Unity/Textures");
         //}
 
-        private static IEnumerable<string> EnumerateAssetFilesAt(string dir, params string[] extensions)
+        private static IEnumerable<string> EnumerateAssetFilesAt(string dir, params string[] endPatterns)
         {
             foreach (string f in Directory.GetFiles(dir))
             {
-                if (extensions.Any(ext => String.Compare(ext, System.IO.Path.GetExtension(f), true) == 0))
+                if (endPatterns.Any(pat => f.EndsWith(pat, true, null)))
                 {
                     yield return f;
                 }
@@ -61,7 +63,7 @@ namespace Tiled2Unity
 
             foreach (string d in Directory.GetDirectories(dir))
             {
-                foreach (string f in EnumerateAssetFilesAt(d, extensions))
+                foreach (string f in EnumerateAssetFilesAt(d, endPatterns))
                 {
                     yield return f;
                 }
@@ -81,6 +83,12 @@ namespace Tiled2Unity
 
                 // Just to be safe. Do not remove scripts.
                 if (f.EndsWith(".cs", true, null))
+                    continue;
+
+                // Do not remove special materials
+                if (f.EndsWith("t2uSprite-Depth.mat", true, null))
+                    continue;
+                if (f.EndsWith("t2uSprite-DiffuseDepth.mat", true, null))
                     continue;
 
                 AssetDatabase.DeleteAsset(f);
